Decompose Corpuses.tsx into thinner composer + sibling files
`frontend/src/views/Corpuses.tsx` had grown to 3,336 lines — listed in #1446
as oversized. Three extractions land it under 1,900:

- `Corpuses.styles.ts` (new): all sidebar / tab-header / mobile-menu
  styled-components used by the main view, plus the tab-header components
  shared with the extracted ExtractsTabContent.
- `CorpusQueryView.tsx` (new): the corpus dashboard / chat composer rendered
  inside the Home tab, together with its private DashboardContainer /
  ChatNavigationHeader / etc. styled-components. Owns the local search-input
  ↔ expanded-chat transition.
- `ExtractsTabContent.tsx` (new): the split-view extracts tab, together with
  its private ExtractsSplitView / ExtractsListPane / ExtractsDetailPane /
  toolbar styled-components.

Net: main file drops from 3,336 → 1,863 lines (44% reduction) while
remaining behaviorally identical — the public `Corpuses` export, route
wiring, and rendered DOM are unchanged. Dead code removed along the way:
- `ChatTransitionContainer`, `SearchToConversationInput`, `NotificationBadge`
  styled-components (declared, never rendered).
- `opened_corpus_id` prop and `handleSearchSubmit` handler on the inner
  query view (never read).
- Imports that became unreferenced after the extraction (`Home` / `Search`
  / `History` icons, `CorpusHome`, `CORPUS_*` design-token re-exports,
  `showQueryViewState`, `selectedExtractIds`, `updateAnnotationSelectionParams`,
  `CorpusExtractCards`, `CorpusExtractDetail`).

State doesn't need to be lifted to a Jotai atom for this slice; the two
extracted components own only local UI state, and everything they need from
the parent is already passed as props.

Refs #1446
